services/grafana/exporters/requirements.txt
prometheus-client==0.19.0
psycopg2-binary==2.9.9

